+++ airbrakes/data_handling/data_processor.py
@@ -7,7 +7,7 @@ import numpy.typing as npt
 from scipy.spatial.transform import Rotation as R
 
 from airbrakes.data_handling.packets.imu_data_packet import EstimatedDataPacket
-from airbrakes.data_handling.packets.processed_data_packet import ProcessedDataPacket
+from airbrakes.data_handling.packets.processor_data_packet import ProcessedDataPacket
 from constants import ACCELERATION_NOISE_THRESHOLD, GRAVITY
 from utils import deadband
