+++ airbrakes/data_handling/logger.py
@@ -9,10 +9,10 @@ from typing import Any, Literal
 
 from msgspec import to_builtins
 
-from airbrakes.data_handling.packets.debug_packet import DebugPacket
+from airbrakes.data_handling.packets.context_data_packet import DebugPacket
 from airbrakes.data_handling.packets.imu_data_packet import EstimatedDataPacket, IMUDataPacket
-from airbrakes.data_handling.packets.logged_data_packet import LoggedDataPacket
-from airbrakes.data_handling.packets.processed_data_packet import ProcessedDataPacket
+from airbrakes.data_handling.packets.logger_data_packet import LoggedDataPacket
+from airbrakes.data_handling.packets.processor_data_packet import ProcessedDataPacket
 from constants import IDLE_LOG_CAPACITY, LOG_BUFFER_SIZE, STOP_SIGNAL
